+++ libraries/compute-utils/function-views/src/fitting/optimizer-nelder-mead.ts
@@ -61,21 +61,21 @@ export const nelderMeadSettingsOpts = new Map<string, Setting>([
 ]);
 
 async function getInitialParams(
-  objectiveFunc: (x: Float32Array) => Promise<number>,
+  objectiveFunc: (x: Float64Array) => Promise<number>,
   settings: Map<string, number>,
-  paramsInitial: Float32Array,
-  restrictionsBottom: Float32Array,
-  restrictionsTop: Float32Array): Promise<[Float32Array[], number[]]> {
+  paramsInitial: Float64Array,
+  restrictionsBottom: Float64Array,
+  restrictionsTop: Float64Array): Promise<[Float64Array[], number[]]> {
   const dim = paramsInitial.length + 1;
   const dimParams = paramsInitial.length;
   const nonZeroParam = settings.get('nonZeroParam')!;
   const initScale = settings.get('initialScale')!;
 
-  const optParams = new Array<Float32Array>(dim);
+  const optParams = new Array<Float64Array>(dim);
   const pointObjectives = new Array<number>(dim);
 
   for (let i = 0; i < dim; i++) {
-    optParams[i] = new Float32Array(dimParams);
+    optParams[i] = new Float64Array(dimParams);
     for (let j = 0; j < dimParams; j++) {
       optParams[i][j] = paramsInitial[j];
       if (i != 0 && i - 1 === j) {
@@ -97,7 +97,7 @@ async function getInitialParams(
   return [optParams, pointObjectives];
 }
 
-function fillCentroid(centroid: Float32Array, dimParams: number, lastIndex: number, optParams: Float32Array[]) {
+function fillCentroid(centroid: Float64Array, dimParams: number, lastIndex: number, optParams: Float64Array[]) {
   for (let i = 0; i < dimParams; i++) {
     let val = 0;
     for (let j = 0; j < dimParams + 1; j++) {
@@ -110,11 +110,11 @@ function fillCentroid(centroid: Float32Array, dimParams: number, lastIndex: numb
 }
 
 function fillPoint(
-  centroid: Float32Array, point: Float32Array,
-  lastIndex: number, optParams: Float32Array[],
+  centroid: Float64Array, point: Float64Array,
+  lastIndex: number, optParams: Float64Array[],
   scale: number, dimParams: number,
-  restrictionsBottom: Float32Array,
-  restrictionsTop: Float32Array) {
+  restrictionsBottom: Float64Array,
+  restrictionsTop: Float64Array) {
   for (let i = 0; i < dimParams; i++) {
     point[i] = centroid[i];
     point[i] += scale * (centroid[i] - optParams[lastIndex][i]);
@@ -127,11 +127,11 @@ function fillPoint(
 }
 
 export const optimizeNM:IOptimizer = async function(
-  objectiveFunc: (x: Float32Array) => Promise<number>,
-  paramsInitial: Float32Array,
+  objectiveFunc: (x: Float64Array) => Promise<number>,
+  paramsInitial: Float64Array,
   settings: Map<string, number>,
-  restrictionsBottom: Float32Array,
-  restrictionsTop: Float32Array,
+  restrictionsBottom: Float64Array,
+  restrictionsTop: Float64Array,
   threshold?: number,
 ) : Promise<Extremum> {
   // Settings initialization
@@ -158,10 +158,10 @@ export const optimizeNM:IOptimizer = async function(
   let previousBest = 0;
   let noImprovment = 0;
 
-  const centroid = new Float32Array(dimParams);
-  const reflectionPoint = new Float32Array(dimParams);
-  const expansionPoint = new Float32Array(dimParams);
-  const contractionPoint = new Float32Array(dimParams);
+  const centroid = new Float64Array(dimParams);
+  const reflectionPoint = new Float64Array(dimParams);
+  const expansionPoint = new Float64Array(dimParams);
+  const contractionPoint = new Float64Array(dimParams);
   const costs = new Array<number>(maxIter);
 
   if (dim > 1) {
